♻️ Refactor: s3에 업로드하는 인프라 로직과 비즈니스 로직이 혼재된 부분을 이미지를 처리하는 서비스로 분리하여 리팩토링
<footer>
#1
