+++ src/main/java/clap/server/application/service/task/CreateTaskService.java
@@ -2,22 +2,17 @@ package clap.server.application.service.task;
 
 import clap.server.adapter.inbound.web.dto.task.request.CreateTaskRequest;
 import clap.server.adapter.inbound.web.dto.task.response.CreateTaskResponse;
-
 import clap.server.adapter.outbound.persistense.entity.member.constant.MemberRole;
 import clap.server.adapter.outbound.persistense.entity.notification.constant.NotificationType;
-import clap.server.application.mapper.AttachmentMapper;
 import clap.server.application.mapper.response.TaskResponseMapper;
 import clap.server.application.port.inbound.domain.CategoryService;
 import clap.server.application.port.inbound.domain.MemberService;
 import clap.server.application.port.inbound.task.CreateTaskUsecase;
-import clap.server.application.port.outbound.s3.S3UploadPort;
-import clap.server.application.port.outbound.task.CommandAttachmentPort;
 import clap.server.application.port.outbound.task.CommandTaskPort;
+import clap.server.application.service.attachment.AttachmentService;
 import clap.server.application.service.webhook.SendNotificationService;
 import clap.server.common.annotation.architecture.ApplicationService;
-import clap.server.common.constants.FilePathConstants;
 import clap.server.domain.model.member.Member;
-import clap.server.domain.model.task.Attachment;
 import clap.server.domain.model.task.Category;
 import clap.server.domain.model.task.Task;
 import lombok.RequiredArgsConstructor;
@@ -34,8 +29,7 @@ public class CreateTaskService implements CreateTaskUsecase {
     private final MemberService memberService;
     private final CategoryService categoryService;
     private final CommandTaskPort commandTaskPort;
-    private final CommandAttachmentPort commandAttachmentPort;
-    private final S3UploadPort s3UploadPort;
+    private final AttachmentService attachmentService;
     private final SendNotificationService sendNotificationService;
 
     @Override
@@ -49,20 +43,13 @@ public class CreateTaskService implements CreateTaskUsecase {
         Task savedTask = commandTaskPort.save(task);
 
         if (files != null) {
-            fileSize = saveAttachments(files, savedTask);
+            fileSize = attachmentService.saveTaskAttachments(savedTask, files);
         }
         savedTask.finalSave(fileSize);
         commandTaskPort.save(savedTask);
 
         publishNotification(savedTask);
         return TaskResponseMapper.toCreateTaskResponse(savedTask);
-    }
-
-    private int saveAttachments(List<MultipartFile> files, Task task) {
-        List<String> fileUrls = s3UploadPort.uploadFiles(FilePathConstants.TASK_FILE, files);
-        List<Attachment> attachments = AttachmentMapper.toTaskAttachments(task, files, fileUrls);
-        commandAttachmentPort.saveAll(attachments);
-        return fileUrls.size();
     }
 
     private void publishNotification(Task task) {
